[Product] Load actual product info when navigating to ProductScreen
- replaces placeholder image and text with product data
  loaded from a hard-coded, local list of products

+++ app/src/main/java/com/example/fakeamazon/features/home/HomeScreen.kt
@@ -42,7 +42,7 @@ import com.example.fakeamazon.features.home.view.TopHomeSection
 fun HomeScreenRoot(
     modifier: Modifier = Modifier,
     innerPadding: PaddingValues = PaddingValues(),
-    onViewProduct: () -> Unit = {},
+    onViewProduct: (Int) -> Unit = {},
     viewModel: HomeViewModel = hiltViewModel()
 ) {
     LaunchedEffect(Unit) {
